Code review edits
Class variables -- formatted according to Google style
Line 97 -- Consistent styling with the brackets
Line 107-125 -- Consistent styling with the brackets. Trimmed
whitespace. I'm biased in this direction--I love Vim, and removing any
linebreaks within a method eases my speed of navigation. Plus, I think
it generally looks better.

In general, I corrected spacing. I followed Google's advice on
whitespace--the less vertically, the better.

Other observations:

You very often phrase your boolean tests in terms of negations, which
somewhat complicates the reading of your code.

For example, on line 193 in the drawSquare method, you have an "if"
statement which is affected by DeMorgan's law--the negation will test
for this: x >= 0 && y >= && x <= kAppWidth && y <= kAppHeight. I'm not
certain on what your intended logic is--I'd assume it's to keep the
cursor on the board and in the array bounds. If so, I've changed the
code to represent that.